Add alert `exn` to Stdlib functions which raise exceptions
Starting with `Stdlib.raise`, `raise_notrace`, and friends. Their alerts
have a ripple effect through the rest of the Stdlib, raising in every
usage of the functions. The functions which raise them are then
annotated with alerts for exceptions which are meant to be bubbled up
(raised transitively), and so on.